+++ extensions/catppuccin/src/search-userstyles.tsx
@@ -1,7 +1,6 @@
-import { useMemo, useState } from "react";
-import { List, ActionPanel, Action, showToast, Toast } from "@raycast/api";
-import { useFetch } from "@raycast/utils";
-import yaml from "js-yaml";
+import { ActionPanel, Action, List } from "@raycast/api";
+import { SearchList } from "./components/SearchList";
+import { isValidUrl } from "./utils/data.util";
 
 interface UserStyleDetails {
   name: string | string[];
@@ -16,71 +15,46 @@ interface UserStyleDetails {
   pastMaintainers?: string[];
 }
 
-interface UserStylesYaml {
-  userstyles: Record<string, UserStyleDetails>;
-}
+const getAppLink = (appLink?: string | string[]) => {
+  if (!appLink) return undefined;
+  return Array.isArray(appLink) ? appLink[0] : appLink;
+};
 
 export default function SearchUserStyles() {
-  const { isLoading, data, revalidate } = useFetch(
-    "https://raw.githubusercontent.com/catppuccin/userstyles/main/scripts/userstyles.yml",
-  );
-  const [searchText, setSearchText] = useState<string>("");
+  const renderItem = (styleKey: string, userStyleDetails: UserStyleDetails) => {
+    const githubLink = `https://github.com/catppuccin/userstyles/tree/main/styles/${styleKey}`;
+    const appLink = getAppLink(userStyleDetails.readme?.["app-link"]);
+    // remove www http https & trailing slashes if present
+    const formattedAppLink = appLink.replace("https://", "").replace("http://", "").replace("www.", "").replace(/\/$/, "");
 
-  const userStyles = useMemo(() => {
-    if (!data) return {};
+    return (
+      <List.Item
+        key={styleKey}
+        title={Array.isArray(userStyleDetails.name) ? userStyleDetails.name.join(" / ") : userStyleDetails.name}
+        subtitle={`Categories: ${userStyleDetails.categories?.join(", ") || "None"}`}
+        accessories={appLink ? [{ text: formattedAppLink }] : undefined}
+        actions={
+          <ActionPanel>
+            <Action.OpenInBrowser url={githubLink} title="Open GitHub" />
+            {appLink && isValidUrl(appLink) && <Action.OpenInBrowser url={appLink} title="Open App Link" />}
+          </ActionPanel>
+        }
+      />
+    );
+  };
 
-    try {
-      const parsedData = yaml.load(data) as UserStylesYaml;
-      return parsedData.userstyles || {};
-    } catch (error) {
-      showToast(Toast.Style.Failure, "Failed to parse YAML", String(error));
-      return {};
-    }
-  }, [data]);
-
-  const filteredUserStyles = useMemo(() => {
-    return Object.entries(userStyles).filter(([_, userStyleDetails]) => {
-      const styleName = Array.isArray(userStyleDetails.name) ? userStyleDetails.name.join(" ") : userStyleDetails.name;
-      return (
-        styleName.toLowerCase().includes(searchText.toLowerCase()) || _.toLowerCase().includes(searchText.toLowerCase())
-      );
-    });
-  }, [userStyles, searchText]);
-
-  const getAppLink = (appLink?: string | string[]) => {
-    return Array.isArray(appLink) ? appLink[0] : appLink || "No app link provided";
+  const filterFunction = (styleKey: string, userStyleDetails: UserStyleDetails, searchText: string) => {
+    const lowerSearchText = searchText.toLowerCase();
+    const styleName = Array.isArray(userStyleDetails.name) ? userStyleDetails.name.join(" ") : userStyleDetails.name;
+    return styleName.toLowerCase().includes(lowerSearchText) || styleKey.toLowerCase().includes(lowerSearchText);
   };
 
   return (
-    <List
-      isLoading={isLoading}
+    <SearchList<UserStyleDetails>
+      dataKey="styles"
       searchBarPlaceholder="Search user styles..."
-      onSearchTextChange={setSearchText}
-      actions={
-        <ActionPanel>
-          <Action title="Reload" onAction={() => revalidate()} />
-        </ActionPanel>
-      }
-    >
-      {filteredUserStyles.map(([styleKey, userStyleDetails]) => {
-        const githubLink = `https://github.com/catppuccin/userstyles/tree/main/styles/${styleKey}`;
-        const appLink = getAppLink(userStyleDetails.readme?.["app-link"]);
-
-        return (
-          <List.Item
-            key={styleKey}
-            title={Array.isArray(userStyleDetails.name) ? userStyleDetails.name.join(" / ") : userStyleDetails.name}
-            subtitle={`Categories: ${userStyleDetails.categories?.join(", ") || "None"}`}
-            accessories={[{ text: appLink }]}
-            actions={
-              <ActionPanel>
-                <Action.OpenInBrowser url={githubLink} title="Open GitHub" />
-                {appLink && <Action.OpenInBrowser url={appLink} title="Open App Link" />}
-              </ActionPanel>
-            }
-          />
-        );
-      })}
-    </List>
+      renderItem={renderItem}
+      filterFunction={filterFunction}
+    />
   );
 }