+++ extensions/catppuccin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Catppuccin Changelog
 
+<<<<<<< Updated upstream
+=======
+## [Stability Changes & Minor Visual Improvements] - {PR_MERGE_DATE}
+
+- Cleaned up application links for userstyles
+- Add fallback data for userstyles and ports
+- Add port categories to list accessories
+
+>>>>>>> Stashed changes
 ## [Initial Version] - 2024-09-11
 
 - Add `search-palette` command